Version 8
Changes primarily in response to feedback from Will, Chris, and Paris.  The parts of the project page having to do with the wranglers was simplified.  Several fields that were similar to but not identical with existing metadata fields converted to match the existing fields. See the 0.8 section of versions.txt for more detail.